+++ src/DBAL/FirebirdBooleanType.php
@@ -4,16 +4,8 @@ declare(strict_types=1);
 
 namespace Satag\DoctrineFirebirdDriver\DBAL;
 
-use Doctrine\DBAL\Platforms\AbstractPlatform;
 use Doctrine\DBAL\Types\BooleanType;
-use Override;
-use Satag\DoctrineFirebirdDriver\Platforms\FirebirdPlatform;
 
 final class FirebirdBooleanType extends BooleanType
 {
-    #[Override]
-    public function requiresSQLCommentHint(AbstractPlatform $platform): bool
-    {
-        return $platform::class === FirebirdPlatform::class;
-    }
 }