+++ src/Platforms/Keywords/FirebirdKeywords.php
@@ -5,22 +5,14 @@ declare(strict_types=1);
 namespace Satag\DoctrineFirebirdDriver\Platforms\Keywords;
 
 use Doctrine\DBAL\Platforms\Keywords\KeywordList;
-use Override;
 
 class FirebirdKeywords extends KeywordList
 {
-    #[Override]
     public function getName(): string
     {
         return 'Firebird';
     }
 
-    /**
-     * {@inheritDoc}
-     *
-     * @link https://firebirdsql.org/refdocs/langrefupd25-reskeywords-full-reswords.html
-     */
-    #[Override]
     protected function getKeywords(): array
     {
         return [